feat(3181): Skip execution of a virtual job when an event is started from that job
