+++ package.json
@@ -118,7 +118,7 @@
     "screwdriver-executor-queue": "^5.0.0",
     "screwdriver-executor-router": "^4.0.0",
     "screwdriver-logger": "^2.0.0",
-    "screwdriver-models": "^30.2.0",
+    "screwdriver-models": "^31.0.0",
     "screwdriver-notifications-email": "^4.0.0",
     "screwdriver-notifications-slack": "^6.0.0",
     "screwdriver-request": "^2.0.1",
